move check functions in DescriptionScrutinizer to CommonDescriptionScrutinizer
